requirements.txt
fastapi 
uvicorn 
langchain 
langgraph 
langchain-community
langchain-text-splitters
langchain-openai
langchain-pinecone
pinecone-client
pypdf
python-multipart
python-dotenv
httpx
tiktoken
sse-starlette
groq
streamlit
requests
sentence-transformers
